- Added background bar for Lives/gold/energy at the top  - Added tower snap to invisible grid (size is 34x34) (towers are 30x30)  - Changed tower sprite to smaller temporary sprite (64x64)  - Can no longer place a tower inside the tower select zone on the right
TODO: Make a grid show up when a tower is selected, need to draw 4pixel lines between squares somehow, might have to hardcode it in.
Not sure how i'll draw the lines, hopefully theres an easy function built in.
I added some code in preparation for drawing mostly invisible squares in each spot,
also added the square's image already too. We gotta figure out how we are going to mark the squares that cant be placed on.
Should be easy to check if a tower has already been placed if the towers are storing their locations. (they are I think?)
As for making it not place in the lane, that'll be a lot harder. Probably once we redraw the map a bit to fit neater in the grid we can figure it out.